+++ apps/api/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.11.1] - 2026-08-09
+
+### Changed
+
+- Upgraded `@phylopic/utils/svg` to `1.2.1`.
+
 ## [2.11.0] - 2026-08-09
 
 ### Changed