+++ packages/utils/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.2.1] - 2026-08-09
+
+### Security
+
+- SVG sanitization now removes Adobe Illustrator `i:pgf` metadata and orphaned base64 text left behind when DOMPurify strips unknown elements.
+
 ## [1.2.0] - 2026-08-09
 
 ### Added